resuse mock struct in service discovery test
