+++ registry/consul/service_discovery.go
@@ -38,7 +38,6 @@ import (
 	"github.com/apache/dubbo-go/common/extension"
 	"github.com/apache/dubbo-go/common/logger"
 	"github.com/apache/dubbo-go/config"
-	_ "github.com/apache/dubbo-go/metadata/mapping/memory"
 	"github.com/apache/dubbo-go/registry"
 )
